requirements.txt
fastapi
uvicorn
aiohttp
openai
duckduckgo_search
python-dotenv
python-multipart
